[FIX] @wlsk0227 회원정보조회 alarm count redis 조회로 구현

See merge request s08-final/S08P31D105!19

+++ BE/majoong/src/main/java/com/example/majoong/user/dto/UserInformationDto.java
@@ -9,7 +9,7 @@ public class UserInformationDto {
     private String phoneNumber;
     private String nickname;
     private String profileImage;
-    private int alarmCount;
+    private long alarmCount;
 
 
 }
